pyproject.toml
[tool.poetry]
name = "kafka-pg-upload"
version = "0.1.0"
description = "Consume data from Kafka topics and upload it into PostgreSQL"
license = "MIT"
authors = ["Pavlo Dmytrenko <mail@pavdmyt.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
confluent-kafka = "^1.5.0"
environs = "^8.0.0"
structlog = "^20.1.0"
asyncpg = "^0.21.0"
python-json-logger = "^0.1.11"
backoff = "^1.10.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
pylint = "^2.6.0"
isort = "^5.4.2"
wrapt = "^1.12.1"
mypy = "^0.782"
psycopg2-binary = "^2.8.5"

[tool.poetry.scripts]
kafka-pg-upload = 'kafka_pg_upload.cmd:run'

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
